cocoa: Now target the 10.5 SDK.
--HG--
extra : convert_revision : svn%3A7afd0f2a-1d33-46b8-9979-f8add37626ee/trunk%40375

+++ cocoa/controllers/MGMassEditionPanel.m
@@ -98,7 +98,7 @@ http://www.hardcoded.net/licenses/hs_license
     return [currencies count];
 }
 
-- (int)comboBox:(NSComboBox *)aComboBox indexOfItemWithStringValue:(NSString *)aString
+- (NSUInteger)comboBox:(NSComboBox *)aComboBox indexOfItemWithStringValue:(NSString *)aString
 {
     aString = [aString lowercaseString];
     for (int i=0; i<[currencies count]; i++)
